+++ packages/featureserver/CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+- geojson validation security vulnerablity
+
 ## [4.0.1] - 08-05-2022
 ### Changed
 - bump winnow version